pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "drupal-devops-copilot"
version = "0.1.0"
description = "AI-powered DevOps Co-Pilot for Drupal: Jira → Plan → Code → Verify → MR"
readme = "README.md"
requires-python = ">=3.10"
authors = [{ name = "Your Team", email = "dev@example.com" }]

dependencies = [
  "requests>=2.31",
  "PyYAML>=6.0",
  "Jinja2>=3.1",
  "python-dotenv>=1.0",
]

[project.optional-dependencies]
# Pick the LLM provider(s) you actually use
llm = [
  "openai>=1.40.0",
]
dev = [
  "pytest>=7.4",
  "ruff>=0.5.0",
  "black>=24.3.0",
  "mypy>=1.10.0",
  "types-requests",
  "types-PyYAML",
]
ui = ["Flask>=3.0"]

# NEW: API + EC2 deploy/QA stacks
api = [
  "fastapi>=0.111",
  "uvicorn[standard]>=0.30",
  "aiohttp>=3.9",
  "pydantic>=2.7",
]
ec2 = [
  "asyncssh>=2.14.2",
  "httpx>=0.27",
]

[project.scripts]
copilot-workflow = "copilot.cli.workflow:main"
copilot-auto-dev = "copilot.cli.auto_dev:main"
copilot-ai-review-merge = "copilot.cli.ai_review_and_merge:main"
copilot-qa-ec2 = "copilot.cli.qa_on_ec2:main"
copilot-one-shot = "copilot.cli.one_shot:main"

[tool.pytest.ini_options]
addopts = "-q"

# ---- setuptools packaging config ----
[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["copilot*"]

# Include non-Python assets (prompts) inside the package
[tool.setuptools.package-data]
"copilot.prompts" = ["*.yaml", "*.jsonl"]
